Plotting detected graph

+++ src/main/java/segmentBlobs/Staticproperties.java
@@ -10,11 +10,11 @@ public final class Staticproperties {
 
 		public final int Label;
 		public final double maxextent;
-		public final int Area;
+		public final double Area;
 		public final double[] location;
 		public final double Intensity;
 		
-		protected Staticproperties(final int Label, final double maxextent, final int Area, final double[] location,
+		protected Staticproperties(final int Label, final double maxextent, final double Area, final double[] location,
 				final double Intensity) {
 			this.Label = Label;
 			this.maxextent = maxextent;
